Moved mqtt_library to a folder
